xmanager_example/requirements.txt
absl-py
pyscf
numpy
